Add popular helper to tracking extension
