+++ ckan/lib/helpers.py
@@ -56,7 +56,7 @@ import ckan
 
 
 from ckan.lib.pagination import Page  # type: ignore # noqa: re-export
-from ckan.common import _, ungettext, g, request, json
+from ckan.common import _, g, request, json
 
 from ckan.lib.webassets_tools import include_asset, render_assets
 from markupsafe import Markup, escape
@@ -1935,23 +1935,6 @@ def remove_url_param(key: Union[list[str], str],
 def debug_inspect(arg: Any) -> Markup:
     ''' Output pprint.pformat view of supplied arg '''
     return literal('<pre>') + pprint.pformat(arg) + literal('</pre>')
-
-
-@core_helper
-def popular(type_: str,
-            number: int,
-            min: int = 1,
-            title: Optional[str] = None) -> str:
-    ''' display a popular icon. '''
-    if type_ == 'views':
-        title = ungettext('{number} view', '{number} views', number)
-    elif type_ == 'recent views':
-        title = ungettext('{number} recent view', '{number} recent views',
-                          number)
-    elif not title:
-        raise Exception('popular() did not recieve a valid type_ or title')
-    return snippet('snippets/popular.html',
-                   title=title, number=number, min=min)
 
 
 @core_helper
