+++ ckanext/tracking/plugin.py
@@ -1,7 +1,7 @@
 import click
 import ckan.plugins as p
 
-from typing import Any
+from typing import Any, Callable
 
 from ckan import model
 from ckan.plugins import toolkit
@@ -9,6 +9,7 @@ from ckan.types import Context, CKANApp
 from ckan.common import CKANConfig
 
 from .cli.tracking import tracking
+from .helpers import popular
 from .middleware import TrackingMiddleware
 
 
@@ -18,6 +19,7 @@ class TrackingPlugin(p.SingletonPlugin):
     p.implements(p.IConfigurer)
     p.implements(p.IMiddleware, inherit=True)
     p.implements(p.IPackageController, inherit=True)
+    p.implements(p.ITemplateHelpers)
 
     # IClick
     def get_commands(self) -> "list[click.Command]":
@@ -50,9 +52,35 @@ class TrackingPlugin(p.SingletonPlugin):
         return pkg_dict
     
 
+    def after_dataset_search(self, search_results: "dict[str, Any]", search_params: "dict[str, Any]") -> "dict[str, Any]":
+        """Add tracking summary to search results.
+        
+        Tracking data is indexed but not stored in Solr so we need to 
+        fetch it from the database. This can cause some discrepancies since 
+        the number of views when indexing might have been different than 
+        when this code is run.
+        """
+        for package_dict in search_results["results"]:
+            tracking_summary = model.TrackingSummary.get_for_package(package_dict["id"])
+            package_dict["tracking_summary"] = tracking_summary
+            for resource_dict in package_dict['resources']:
+                summary =  model.TrackingSummary.get_for_resource(
+                    resource_dict['url']
+                    )
+                resource_dict['tracking_summary'] = summary
+        return search_results
+    
+
     def before_dataset_index(self, pkg_dict: "dict[str, Any]") -> "dict[str, Any]":
         tracking_summary = model.TrackingSummary.get_for_package(pkg_dict["id"])
         pkg_dict['views_total'] = tracking_summary['total']
         pkg_dict['views_recent'] = tracking_summary['recent']
         return pkg_dict
+    
+
+    # ITemplateHelpers
+    def get_helpers(self) -> "dict[str, Callable[...,Any]]":
+        return {
+            "popular": popular,
+        }
